telegram controller minor refactoring
